+++ WalletWasabi.Fluent/Converters/FeeTargetTimeConverter.cs
@@ -1,36 +1,9 @@
 using Avalonia.Data.Converters;
 using System.Globalization;
 using Avalonia;
-using Avalonia.Media;
 using WalletWasabi.Helpers;
 
 namespace WalletWasabi.Fluent.Converters;
-
-public class BoolToBrushConverter : IValueConverter
-{
-	public static BoolToBrushConverter Instance { get; } = new();
-
-	public IBrush? TrueBrush { get; set; }
-
-	public IBrush? FalseBrush { get; set; }
-
-	public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
-	{
-		if (value is true)
-		{
-			return TrueBrush;
-		}
-		else
-		{
-			return FalseBrush;
-		}
-	}
-
-	public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-	{
-		throw new NotImplementedException();
-	}
-}
 
 public class FeeTargetTimeConverter : IValueConverter
 {
